pyproject.toml
[project]
name = "olx-car-parser"
version = "0.1.0"
description = "OLX private car listings parser with price analytics"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "sqlalchemy>=2.0",
    "pydantic>=2.0",
    "pandas>=2.0",
    "matplotlib>=3.8",
    "typer>=0.12",
    "pyyaml>=6.0",
    "apscheduler>=3.10",
    "rich>=13.0",
    "streamlit>=1.38",
    "plotly>=5.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
olx-parser = "src.cli:app"

[tool.pytest.ini_options]
markers = ["smoke: live API health checks"]

[tool.ruff]
target-version = "py311"
line-length = 100
